requirements_frozen_v3.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
ale-py==0.11.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
array-api-compat==1.14.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
bottle==0.13.4
Box2D==2.3.10
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
coverage==7.13.5
cryptography==46.0.6
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.8.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
dill==0.3.8
dm-env==1.6
dm-tree==0.1.10
dm_control==1.0.38
docutils==0.22.4
einops==0.8.2
etils==1.14.0
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.25.2
flax==0.12.6
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gdown==5.2.1
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
griffelib==2.0.2
gymnasium==1.2.3
h11==0.16.0
h5py==3.16.0
hdf5plugin==6.0.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
humanize==4.15.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
id==1.6.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
ipdb==0.13.13
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jax==0.9.2
jaxlib==0.9.2
jaxtyping==0.3.9
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.3
keyring==25.7.0
kiwisolver==1.5.0
labmaze==1.0.6
lightning==2.6.1
lightning-utilities==0.15.3
loguru==0.7.3
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mergedeep==1.3.4
minari==0.5.3
minigrid==3.0.0
mkdocs==1.6.1
mkdocs-autorefs==1.4.4
mkdocs-get-deps==0.2.2
mkdocs-shadcn==0.10.3
mkdocstrings==1.0.3
mkdocstrings-python==2.0.3
ml_dtypes==0.5.4
more-itertools==11.0.1
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.6.0
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.6.1
nh3==0.3.4
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.19.0.56
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.28.9
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.8.90
ogbench==1.2.1
omegaconf==2.3.0
opencv-python==4.13.0.92
opt_einsum==3.4.0
optax==0.2.8
orbax-checkpoint==0.11.33
orbax-export==0.0.8
packaging==26.0
pandas==3.0.2
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prettytable==3.17.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.20.0
pymdown-extensions==10.21.2
pymunk==7.2.0
PyOpenGL==3.1.10
pyparsing==3.3.2
PySocks==1.7.1
pytest==9.0.2
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytorch-lightning==2.6.1
PyYAML==6.0.3
pyyaml_env_tag==1.1
readme_renderer==44.0
regex==2026.4.4
requests==2.33.1
requests-cache==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
richuru==0.1.1
ruff==0.15.9
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.5.0
sentry-sdk==2.57.0
setuptools==70.2.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stable-pretraining==0.1.6
-e git+ssh://git@github.com/epokhcs/stable-worldmodel.git@606959c25fba9c690114fb5625cdd433d99a21a7#egg=stable_worldmodel
stable_baselines3==2.8.0
stack-data==0.6.3
submitit==1.5.4
swig==4.4.1
sympy==1.14.0
tabulate==0.10.0
tensorstore==0.1.82
threadpoolctl==3.6.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0+cu128
torchmetrics==1.9.0
torchvision==0.26.0+cu128
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.0
treescope==0.1.10
triton==3.6.0
twine==6.2.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
url-normalize==2.2.1
urllib3==2.6.3
uvloop==0.22.1
wadler_lindig==0.1.7
wandb==0.25.1
watchdog==6.0.0
wcwidth==0.6.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
